initial implementation of cli tool
